# accounts/abi: Improve unpack performance