Reduces alloc by extra loop iteration, slight performance improvement on my local machine
```
            │   MASTER    │                 PR                 │
            │   sec/op    │   sec/op     vs base               │
Unpack/0-32   217.9n ± 5%   170.3n ± 2%  -21.84% (p=0.002 n=6)
Unpack/1-32   628.7n ± 9%   573.5n ± 7%   -8.78% (p=0.002 n=6)
```