@MariusVanDerWijden 2025-03-14
Sorry that it took so long, I was unable to push to your branch, so I opened a new PR: https://github.com/ethereum/go-ethereum/pull/31387
I will close this one for now. You will get full attribution once the PR is merged into master, thanks for your contribution!